+++ __mocks__/@grafana/ui.tsx
@@ -40,6 +40,14 @@ const MultiSelect = forwardRef((props, ref) => <Select {...props} ref={ref} mult
 
 MultiSelect.displayName = 'MultiSelect';
 
+// Using this specifically for Probe filter at the moment
+const mockProbeOptions = [{ label: 'Chicago', value: 22 }];
+const AsyncMultiSelect = forwardRef((props, ref) => (
+  <Select {...props} options={mockProbeOptions} ref={ref} multiple={true} />
+));
+
+AsyncMultiSelect.displayName = 'AsyncMultiSelect';
+
 const Icon = (props) => <svg {...props} />;
 
 module.exports = {
@@ -47,4 +55,5 @@ module.exports = {
   MultiSelect,
   Select,
   Icon,
+  AsyncMultiSelect,
 };