+++ src/dashboards/loader.ts
@@ -46,7 +46,7 @@ export async function importDashboard(
   const json = await backendSrv.request({
     url: `public/plugins/grafana-synthetic-monitoring-app/dashboards/${path}`,
     method: 'GET',
-    headers: { 'Cache-Control': 'no-cache' },
+    headers: { 'Cache-Control': 'no-store' },
   });
 
   const folder = await findSyntheticMonitoringFolder();
@@ -79,7 +79,7 @@ export async function listAppDashboards(): Promise<DashboardInfo[]> {
     const json = await backendSrv.request({
       url: `public/plugins/grafana-synthetic-monitoring-app/dashboards/${p}`,
       method: 'GET',
-      headers: { 'Cache-Control': 'no-cache' },
+      headers: { 'Cache-Control': 'no-store' },
     });
 
     const dInfo = {